refactor(check): move find command to check flag
